debugged and improved the examples

+++ nipy/labs/utils/reproducibility_measures.py
@@ -675,7 +675,7 @@ def group_reproducibility_metrics(
     # compute the group mask
     affine = load(mask_images[0]).get_affine()
     mask = intersect_masks(mask_images, threshold=0) > 0
-    grp_mask = Nifti1Image(mask, affine)
+    grp_mask = Nifti1Image(mask.astype('u8'), affine)
     xyz = np.where(mask)
     xyz = np.array(xyz).T
